Implemented selector dialog feature
We can now select some experiments among the filtered results by
choosing a set of properties unique to the experiment. This allows us
to search experiments hierarchically for example and compare few of
the experiments . This new feature basically makes it easier to select a
bunch of experiments through filtering instead of manually selecting
experiments.

The css files used for styling the dialog box have to be served from the
static url. This would have to be configured by the user and the static
files have to be collected by running manage.py collectstatic