Stored dir props in HTML5 data attributes
The unique properties corresponding to each directory are now added to
a custom data attribute 'data-expsift-uniq-props' of the select
checkboxes. This field has fewer restrictions on what can be stored in
it compared to the class name. This allows us to match symbols like
equalto etc. This is also a cleaner approach than overloading the class
name.